Advancing

+++ crates/rdp/rdp/src/geom.rs
@@ -30,27 +30,27 @@
 // Authors: Adolfo Gómez, dkmaster at dkmon dot com
 #[derive(Debug, Copy, Clone)]
 pub struct Rect {
-    pub x: u32,
-    pub y: u32,
+    pub x: i32,
+    pub y: i32,
     pub w: u32,
     pub h: u32,
 }
 
 impl Rect {
-    pub fn new(x: u32, y: u32, w: u32, h: u32) -> Self {
+    pub fn new(x: i32, y: i32, w: u32, h: u32) -> Self {
         Self { x, y, w, h }
     }
 
     pub fn union(&self, other: &Rect) -> Rect {
         let x1 = self.x.min(other.x);
         let y1 = self.y.min(other.y);
-        let x2 = (self.x + self.w).max(other.x + other.w);
-        let y2 = (self.y + self.h).max(other.y + other.h);
+        let x2 = (self.x + self.w as i32).max(other.x + other.w as i32);
+        let y2 = (self.y + self.h as i32).max(other.y + other.h as i32);
         Rect {
             x: x1,
             y: y1,
-            w: x2 - x1,
-            h: y2 - y1,
+            w: (x2 - x1) as u32,
+            h: (y2 - y1) as u32,
         }
     }
 }
@@ -58,8 +58,8 @@ impl Rect {
 impl From<&freerdp_sys::GDI_RGN> for Rect {
     fn from(rgn: &freerdp_sys::GDI_RGN) -> Self {
         Self {
-            x: rgn.x as u32,
-            y: rgn.y as u32,
+            x: rgn.x as i32,
+            y: rgn.y as i32,
             w: rgn.w as u32,
             h: rgn.h as u32,
         }
